Reorganize src/utils (#130)
Reorganizes the src/utils folder to provide a more coherent hierarchy. Renames many of the utils files and cleans up import order along the way.
