+++ src/components/charts/program_evaluation/us_nd/free_through_recovery/FtrReferralCountByMonth.js
@@ -16,12 +16,12 @@
 // =============================================================================
 
 import React, { useState, useEffect } from 'react';
-
 import { Line } from 'react-chartjs-2';
-import { configureDownloadButtons } from '../../../../../assets/scripts/utils/downloads';
+
 import { COLORS } from '../../../../../assets/scripts/constants/colors';
-import { monthNamesWithYearsFromNumbers } from '../../../../../utils/monthConversion';
-import { sortFilterAndSupplementMostRecentMonths } from '../../../../../utils/dataOrganizing';
+import { configureDownloadButtons } from '../../../../../assets/scripts/utils/downloads';
+import { sortFilterAndSupplementMostRecentMonths } from '../../../../../utils/transforms/datasets';
+import { monthNamesWithYearsFromNumbers } from '../../../../../utils/transforms/months';
 
 const FtrReferralCountByMonth = (props) => {
   const [chartLabels, setChartLabels] = useState([]);
