+++ src/components/charts/program_evaluation/us_nd/free_through_recovery/FtrReferralsByAge.js
@@ -16,11 +16,11 @@
 // =============================================================================
 
 import React, { useState, useEffect } from 'react';
-
 import { Bar } from 'react-chartjs-2';
+
 import { COLORS } from '../../../../../assets/scripts/constants/colors';
 import { configureDownloadButtons } from '../../../../../assets/scripts/utils/downloads';
-import { toInt } from '../../../../../utils/variableConversion';
+import { toInt } from '../../../../../utils/transforms/labels';
 
 const FtrReferralsByAge = (props) => {
   const [chartLabels, setChartLabels] = useState([]);
