+++ src/components/charts/revocations/RevocationsByOffice.js
@@ -27,10 +27,11 @@ import {
 import ReactTooltip from 'react-tooltip';
 import { geoAlbersUsa } from 'd3-geo';
 import { scaleLinear } from 'd3-scale';
-import geographyObject from '../../../assets/static/maps/us_nd.json';
+
 import { COLORS } from '../../../assets/scripts/constants/colors';
 import { configureDownloadButtons } from '../../../assets/scripts/utils/downloads';
-import { toHtmlFriendly, toInt } from '../../../utils/variableConversion';
+import geographyObject from '../../../assets/static/maps/us_nd.json';
+import { toHtmlFriendly, toInt } from '../../../utils/transforms/labels';
 
 const chartId = 'revocationsByOffice';
 const centerNDLong = -100.5;
