+++ src/components/charts/snapshots/SupervisionSuccessSnapshot.js
@@ -16,18 +16,20 @@
 // =============================================================================
 
 import React, { useState, useEffect } from 'react';
-
 import { Line } from 'react-chartjs-2';
-import { configureDownloadButtons } from '../../../assets/scripts/utils/downloads';
+
 import { COLORS } from '../../../assets/scripts/constants/colors';
-import { monthNamesWithYearsFromNumbers } from '../../../utils/monthConversion';
-import { sortAndFilterMostRecentMonths } from '../../../utils/dataOrganizing';
-import { generateTrendlineDataset, getTooltipWithoutTrendline } from '../../../utils/trendline';
+import { configureDownloadButtons } from '../../../assets/scripts/utils/downloads';
+import { toInt } from '../../../utils/transforms/labels';
+import { sortAndFilterMostRecentMonths } from '../../../utils/transforms/datasets';
+import { monthNamesWithYearsFromNumbers } from '../../../utils/transforms/months';
 import {
   getGoalForChart, getMinForGoalAndData, getMaxForGoalAndData, trendlineGoalText,
   goalLabelContentString,
-} from '../../../utils/metricGoal';
-import { toInt } from '../../../utils/variableConversion';
+} from '../../../utils/charts/metricGoal';
+import {
+  generateTrendlineDataset, getTooltipWithoutTrendline,
+} from '../../../utils/charts/trendline';
 
 const SupervisionSuccessSnapshot = (props) => {
   const [chartLabels, setChartLabels] = useState([]);
