+++ src/utils/authentication/viewAuthentication.js
